apk: Embed configurable assets in `asset/` directory
Android apps might want to ship with extra files - assets - to
complement the native build.  These could be single files or folders
that are included recursively.  the `file_name()` of the given path is
added directly to the `assets/` directory, and any directory structure
beyond that point is maintained.

Assets are specified in the following way in `manifest.yaml`:

    android:
      assets:
        - some/path
        - { path: some/path } # Same as above
        - { path: some/path, optional: true } # Do not error if this  path doesn't exist
        - { path: some/path, alignment: 4096 } # Page-aligned for optimal use with AASSET_MODE_BUFFER and AAsset_getBuffer
        - { path: some/path, alignment: unaligned }
        - { path: some/path, alignment: compressed }